[refactor]Rename Method
FindRestaurant 패키지
EatingPlace->Restaurant
